Add `Expr::get_type()`.
This function, in most cases, returns inferrable atomic types based
on the kind of expression, but at times has to reference symbols
or stored state. This is for convenience when working with expressions
while constructing the LLVM IR.

Some changes are also made to `typecheck.rs` making expression-visiting
methods take mutable references to expressions in order to set the type
of array initializers and index expressions.